+++ Pipfile
@@ -16,6 +16,7 @@ flask-pymongo = "*"
 jose = "*"
 datetime = "*"
 uuid = "*"
+mongoengine = "*"
 
 [dev-packages]
