+++ src/routes/__init__.py
@@ -1,0 +1,1 @@
+from src.routes.users import user_blueprint
